build(Javadoc): removed some dependencies and move class
Remove GraphQL dependency and endpoint (was blocking Javadoc generation)
Move WorkerLifeCycle into package to resolve Javadoc name clash error
