+++ pom.xml
@@ -292,15 +292,6 @@
             <name>Open Source Geospatial Foundation Repository</name>
             <url>https://download.osgeo.org/webdav/geotools/</url>
         </repository>
-        <!-- Needed  for graphql-java-->
-        <repository>
-            <snapshots>
-                <enabled>false</enabled>
-            </snapshots>
-            <id>bintray-andimarek-graphql-java</id>
-            <name>bintray</name>
-            <url>https://dl.bintray.com/andimarek/graphql-java</url>
-        </repository>
     </repositories>
 
     <dependencies>
@@ -411,11 +402,6 @@
             <groupId>com.conveyal</groupId>
             <artifactId>kryo-tools</artifactId>
             <version>1.3.0</version>
-        </dependency>
-        <dependency>
-            <groupId>com.graphql-java</groupId>
-            <artifactId>graphql-java</artifactId>
-            <version>2016-02-19T11-51-00</version>
         </dependency>
         <dependency>
             <groupId>com.sparkjava</groupId>
