+++ src/main/java/com/conveyal/r5/otp2/rangeraptor/debug/DebugHandlerFactory.java
@@ -4,7 +4,7 @@ import com.conveyal.r5.otp2.api.path.Path;
 import com.conveyal.r5.otp2.api.request.DebugRequest;
 import com.conveyal.r5.otp2.api.transit.TripScheduleInfo;
 import com.conveyal.r5.otp2.api.view.ArrivalView;
-import com.conveyal.r5.otp2.rangeraptor.WorkerLifeCycle;
+import com.conveyal.r5.otp2.rangeraptor.workerlifecycle.WorkerLifeCycle;
 import com.conveyal.r5.otp2.rangeraptor.view.DebugHandler;
 import com.conveyal.r5.otp2.util.paretoset.ParetoSetEventListener;
