+++ src/main/java/com/conveyal/r5/otp2/rangeraptor/debug/DebugHandlerStopArrivalAdapter.java
@@ -3,7 +3,7 @@ package com.conveyal.r5.otp2.rangeraptor.debug;
 import com.conveyal.r5.otp2.api.request.DebugRequest;
 import com.conveyal.r5.otp2.api.transit.TripScheduleInfo;
 import com.conveyal.r5.otp2.api.view.ArrivalView;
-import com.conveyal.r5.otp2.rangeraptor.WorkerLifeCycle;
+import com.conveyal.r5.otp2.rangeraptor.workerlifecycle.WorkerLifeCycle;
 
 import java.util.LinkedList;
