+++ src/main/java/com/conveyal/r5/otp2/rangeraptor/multicriteria/McRangeRaptorWorkerState.java
@@ -4,7 +4,7 @@ import com.conveyal.r5.otp2.api.path.Path;
 import com.conveyal.r5.otp2.api.transit.IntIterator;
 import com.conveyal.r5.otp2.api.transit.TransferLeg;
 import com.conveyal.r5.otp2.api.transit.TripScheduleInfo;
-import com.conveyal.r5.otp2.rangeraptor.WorkerLifeCycle;
+import com.conveyal.r5.otp2.rangeraptor.workerlifecycle.WorkerLifeCycle;
 import com.conveyal.r5.otp2.rangeraptor.WorkerState;
 import com.conveyal.r5.otp2.rangeraptor.multicriteria.arrivals.AbstractStopArrival;
 import com.conveyal.r5.otp2.rangeraptor.multicriteria.arrivals.AccessStopArrival;
