+++ src/main/java/com/conveyal/r5/otp2/rangeraptor/transit/SearchContext.java
@@ -11,7 +11,7 @@ import com.conveyal.r5.otp2.api.transit.TransferLeg;
 import com.conveyal.r5.otp2.api.transit.TransitDataProvider;
 import com.conveyal.r5.otp2.api.transit.TripScheduleInfo;
 import com.conveyal.r5.otp2.rangeraptor.RoundProvider;
-import com.conveyal.r5.otp2.rangeraptor.WorkerLifeCycle;
+import com.conveyal.r5.otp2.rangeraptor.workerlifecycle.WorkerLifeCycle;
 import com.conveyal.r5.otp2.rangeraptor.debug.DebugHandlerFactory;
 import com.conveyal.r5.otp2.rangeraptor.debug.WorkerPerformanceTimers;
 import com.conveyal.r5.otp2.rangeraptor.workerlifecycle.LifeCycleBuilder;
